Split out progress view
